Split the monolithic smoke test into smaller ones
